requirements.txt
dvc[s3]
scikit-learn==1.3.2
pandas==2.0.3
numpy==1.24.3
pyyaml
joblib
